requirements.txt
# Root-level requirements for all agents
google-adk>=1.17.0
openai>=1.0.0  # Required for DALL-E image generation in linkedin_content_agent
python-dotenv>=1.0.0  # For loading .env files
pytest>=8.4.2  # For testing blogger agent
pytest-asyncio>=1.2.0
deprecated
pandas
tabulate
todd
scikit-learn
# Customer service agent dependencies
pydantic-settings>=2.8.1  # Required for customer service agent config
cloudpickle>=3.1.1  # Required for customer service agent
jsonschema>=4.23.0  # Required for customer service agent
twilio
numpy
soxr
uvicorn
boto3